+++ test/test_hmfisofox.py
@@ -1,19 +1,19 @@
-import qcetl.hmfbamtools
+import qcetl.hmfisofox
 import test.cachechecker
 
 
-def tests_hmfbamtools():
+def tests_hmfisofox():
     test.cachechecker.check(
-        qcetl.hmfbamtools.HmfBamToolsCache(),
+        qcetl.hmfisofox.HmfIsofoxCache(),
         [
             {
-                "path": "test/files/hmfbamtools/COLO829TESTT.bam_metric.summary.tsv",
-                "swid": "SWID",
+                "path": "test/files/hmfisofox/COLO829TESTR-RNA.isf.summary.csv",
+                "swid": "SWID_RNA",
                 "project": "COLO829TEST",
-                "pinery_lims_ids": ["ID1", "ID2"],
+                "pinery_lims_ids": ["ID3"],
                 "donor": "COLO829TESTT",
                 "group_id": "NO",
-                "library_design": "WG",
+                "library_design": "WT",
                 "tissue_origin": "Sk",
                 "tissue_type": "M",
                 "reference": "hg38",
@@ -21,6 +21,6 @@ def tests_hmfbamtools():
             }
         ],
         {
-            "summary": "test/files/hmfbamtools/hmfbamtools_summary.csv",
+            "summary": "test/files/hmfisofox/hmfisofox_summary.csv",
         },
     )